+++ resources/views/display-seances.blade.php
@@ -50,7 +50,7 @@
                                     <td>{{ $seance->heure_debut }}</td>
                                     <!-- <td>00:00:00</td> -->
                                     <td><a href="/update/{{ $seance->id }}"><img src="../../../pen-edit.svg" class="pen-edit"></a></td>
-                                    <td><a href="/delete/{{ $seance->id }}">&#128465;</a></td>
+                                    <td><a href="/delete/{{ $seance->id }}" onclick="return confirm('Êtes vous sûr de vouloir supprimer cette séance ?');">&#128465;</a></td>
                                 </tr>
                             @endforeach
                         </tbody>
